+++ cardano-db-sync/src/Cardano/DbSync/Era/Shelley/Genesis.hs
@@ -14,6 +14,7 @@ module Cardano.DbSync.Era.Shelley.Genesis (
 import Cardano.BM.Trace (Trace, logError, logInfo)
 import qualified Cardano.Db as DB
 import Cardano.DbSync.Api
+import Cardano.DbSync.Api.Types (SyncEnv)
 import Cardano.DbSync.Cache
 import qualified Cardano.DbSync.Era.Shelley.Generic.Util as Generic
 import Cardano.DbSync.Era.Shelley.Insert
